backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
supabase==2.6.0
python-dotenv==1.0.1
PyJWT==2.9.0
APScheduler==3.10.4
pydantic==2.8.2
pydantic-settings==2.4.0
